travel-advisor-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymongo==4.6.1
chromadb==0.4.22

# LangChain
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5

# ML/AI
sentence-transformers==2.3.1
openai==1.10.0
numpy>=1.24.0
pandas>=2.0.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
torch>=2.0.0

# NLP & Text Processing
unidecode>=1.3.6
rapidfuzz>=3.0.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0

# Background tasks
apscheduler==3.10.4

# Monitoring
prometheus-fastapi-instrumentator==6.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
mypy==1.8.0
